requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
Pillow==6.0.0
pymarc==3.1.13
PyPDF2==1.26.0
requests==2.22.0
six==1.12.0
urllib3==1.25.3